+++ src/main.cpp
@@ -430,6 +430,7 @@ void MyApp::ReInitLanguage()
     else
         MyLogMessage(_("Systemlanguage: the default system language"));
 
+    score::InitTexts4Translation(true);   // force re-init of static texts for translation
 }   // ReInitLanguage()
 
 bool MyApp::OnInit()
@@ -929,7 +930,7 @@ void MyFrame::OnImportSchema(wxCommandEvent& )
         {
             wxString error(_("Don't import a schema from the basefolder itself!"));
             MyLogError(error);
-            MyMessageBox(error, _("ERROR"));
+            MyMessageBox(error, _("Error"));
         } else
         {   // only import if folders are different!
             if (schema::ImportSchema(schemaFile, true))